whisper-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai-whisper==20231117
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
numpy==1.24.3
ffmpeg-python==0.2.0
gunicorn 